[db-label] Only label info elements without titles
In DocBook 5, it is possible to have an info element almost anywhere, and
in the vast majority of cases titles can appear inside the info element instead
of outside.  DocBook 4 always required the title to appear outside the *info
element.  If a DocBook 5 source file has the title inside the info element, it
will be handled correctly by other code.  Therefore, only do automatic labelling
of info elements that do not have a title inside.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.ath.cx>