mal-link: Change how sort titles are found in link getters
Sort titles are no longer required to be in the cache. That's
good, because having them in the cache meant that we couldn't
determine if the page actually defined a sort title, and thus
we couldn't do smarter fallback for different link types. It
was also needless complexity for cache generators.

With this change, we can now use the correct link title to
sort when a sort title is not explicitly defined.

https://bugzilla.gnome.org/show_bug.cgi?id=652126