dita2html-block: Common templates, conref, section titles
Funny story: A section in DITA may optionally have a title.
But you can put that title anywhere in the child list. And
you can have multiple titles. But only the first title gets
used, because there's only supposed to be one, even though
the DTD allows many, anywhere. Just crazy.

Also, it would be nice if the spec would spell this oddity
out very clearly. It doesn't. The only way to be sure is
to assume the DITA-OT is correct.