Allow external info links for topic links
This is the first step in allowing external info links:

https://github.com/projectmallard/projectmallard.org/issues/30

This commit only adds them for topic links, but it adds all the
machinery necessary for other link types to use external links.
The various templates for outputting link content can now all
take an additional info parameter, which is provided by the
templates that list info links.

This commit also adds templates for outputting the desc, which
was previously done ad-hoc in each place desc is used for links.
This not only helps this use, but it will be necessary if we add
multiple descs with roles in a future version of Mallard.