Dropping db.chunk.chunks param in favor of selector entity
Exposing a parameter with a list of elements to chunk was never a
good idea. It gave the impression that you could change the param
and things would actually work. They wouldn't.

But it's nice to have the list of chunkable elements in one place,
so I went with a selector XML entity like I did for DITA.