Do max-width and margin:auto with finer granularity
Previously, the entire article had max-width set with auto left/right
margins. If you wanted to make a section have a wide decorative background
that went all the way to the page edges, you had to override this in your
customization, then set max-width and margin yourself on various divs that
aren't immediately obvious.

This change sets those properties on the not-immediately-obvious divs by
default. It's easier to get it right when we maintain that list upstream.
Wide section backgrounds are popular. With this change, it is literally
as easy as this:

section#mycoolstuff {
  background: url(mycoolstuffbg.png);
}