Use pagewide class to set max content width
By default, we max the content at 960px and center horizontally. To
allow various types of styling, we set this on certain nested divs
and other blocks, instead of on the main article element. With this
change, we no longer use intricate CSS selectors that are tricky to
override. Instead, we output add a pagewide class to those elements.

Don't like a max width?

.pagewide { max-width: 100%; }

Don't like horizontal centering?

.pagewide { margin-left: 0; margin-right: 0; }