Add html.head.top.custom stub template
Sometimes you want to link to a set of common CSS that gets used
across multiple sites, but you want the Yelp stylesheet to take
precedence. Linking in html.head.custom is wrong, becaause that
will override Yelp's rules. This allows you to put stuff at the
to of the head element.