Group UI overlays with JS instead of XSLT
UI overlays don't have a grouping element, but we want to group them
in the HTML so we can use flexbox for row wrap. We were doing the
grouping in XSLT, but this only worked if overlays were exactly next
to each other in document order. If you put them in conditionals,
you lost grouping. Instead, just output them as-is, then use JS to
group them after the fact.