Update highlight.js to 10.5+ and add SML highlighter
Fixes https://gitlab.gnome.org/GNOME/yelp-xsl/-/issues/29

highlight.js renamed cs->csharp and tex->latex, so I had to do the same.
This means that the highlight.js included with yelp-xsl as of this commit
might not highlight those two languages if you drop it into html built
with yelp-xsl before this commit. That's a thing that could potentially
happen in multiple-document sites with partial rebuilds.