common/color: Use different target contrasts for white-on-black
We can compute most colors from a set of base colors by moving colors
closer to the text or background color until they hit a target contrast.
For background colors, we were using a very high contrast with the text
color, because the lighter background colors look nice. But when using
inverse colors (white-on-near-black), that caused the background color
computation to give up from recursion depth and return something far
too dark.

Instead, we now detect if it's an inverse color scheme and adjust the
target contrasts accordingly.

This commit addresses one point in this yelp metabug:
https://gitlab.gnome.org/GNOME/yelp/-/issues/178